requirements.txt
aniso8601==9.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
Werkzeug==2.2.2
